form-controls

+++ app/Http/Controllers/AboutController.php
@@ -1,17 +1,20 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Support\Facades\Log;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
 use App\Models\AboutModel;
+
 
 class AboutController extends Controller
 {
     public function getaboutdata(){ // About tablosundaki verileri çeker ve about viewa gönderir 
       $aboutdata = AboutModel::latest()
                               ->first();
-        
+      if($aboutdata == null){
+        return redirect();
+      }
       return view("about")->with("aboutdata",$aboutdata);
     }
 
@@ -41,13 +44,18 @@ class AboutController extends Controller
     }
     
     public function deleteabout($id){
-      $about = AboutModel::where("id",$id)
-                                ->first();
-      if ($about != null) {
-        $about->delete();
+      $about = AboutModel::where("id", $id)->first();
+      
+      if ($about != null && AboutModel::count() > 1) {
+          $about->delete();
+      } else {
+          
+          return redirect('admin-panel/admin-about-list')->with('error', 'Kayıt silinemedi çünkü yalnızca 1 kayıt kaldı.');
       }
-      return redirect('admin-panel/admin-about-list');
-    }
+  
+      return redirect('admin-panel/admin-about-list')->with('success', 'Kayıt başarıyla silindi.');
+  }
+  
 
     public function updateabout(Request $request){
       $about = AboutModel::where("id",$request->id)
